@@ -3,7 +3,7 @@
     "@types/prop-types": "^15.7.4",
     "@types/react": "^17.0.15",
     "@types/react-dom": "^17.0.9",
-    "typescript-dev-kit": "^3.1.2"
+    "typescript-dev-kit": "^5.0.4"
   },
   "eslintConfig": {
     "extends": [
@@ -11,29 +11,21 @@
     ]
   },
   "dependencies": {
-    "basx": "^1.3.5",
-    "diox": "^4.0.3",
+    "basx": "^2.0.1",
+    "diox": "^5.0.2",
     "prop-types": "^15.7.2",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
-    "sonar-ui": "file:../../../library/dist",
-    "vue": "^2.6.12"
+    "biuty": "file:../../../library/dist",
+    "svelte": "^3.46.4",
+    "vue": "^3.2.31"
   },
   "tsDevKitConfig": {
     "target": "web",
-    "entry": {
-      "react": "./scripts/react.tsx",
-      "vue": "./scripts/vue.ts"
-    },
+    "html": "./html/index.html",
     "devServer": {
-      "ip": "0.0.0.0",
+      "host": "0.0.0.0",
       "port": 5040
-    },
-    "html": {
-      "template": "./html/index.html",
-      "entries": [
-        "react"
-      ]
     },
     "splitChunks": false,
     "srcPath": "src",
@@ -55,7 +47,8 @@
     "dev": "cd node_modules/typescript-dev-kit/ && node scripts/dev",
     "build": "cd node_modules/typescript-dev-kit/ && node scripts/build",
     "test": "cd node_modules/typescript-dev-kit/ && node scripts/test",
-    "doc": "node_modules/.bin/typedoc --out ./doc/ --exclude \"**/*.js\" --exclude \"**/__+(tests|mocks)__/**\" src/",
+    "check": "cd node_modules/typescript-dev-kit/ && node scripts/check",
+    "doc": "node_modules/.bin/typedoc --out ./doc/ --exclude \"**/*.js\" --exclude \"**/__+(tests|mocks)__/**\" src/scripts/react.tsx",
     "postinstall": "rm -f node_modules/.eslintcache"
   }
 }
