@@ -1,5 +1,5 @@
 /**
- * Copyright (c) Matthieu Jabbour. All Rights Reserved.
+ * Copyright (c) Openizr. All Rights Reserved.
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
@@ -15,7 +15,7 @@
  *
  * @returns {string} Generated modifiers list.
  */
-export default function buildClass(baseClass: string, modifiers: string = ''): string {
+export default function buildClass(baseClass: string, modifiers = ''): string {
   const chainedModifiers = [...new Set(modifiers.split(' '))].map((modifier) => (
     (modifier === '') ? '' : `--${modifier}`)).join('');
   return `${baseClass}${(chainedModifiers !== '') ? ` ${baseClass}${chainedModifiers}` : ''}`;
