@@ -1,13 +1,11 @@
 /**
- * Copyright (c) Matthieu Jabbour. All Rights Reserved.
+ * Copyright (c) Openizr. All Rights Reserved.
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  *
  */
 
-declare module '*.vue' {
-  import Vue from 'vue';
+declare module '*.vue';
 
-  export default Vue;
-}
+type JSXElement = any; // eslint-disable-line @typescript-eslint/no-explicit-any
