@@ -1,40 +1,36 @@
 /**
- * Copyright (c) Matthieu Jabbour. All Rights Reserved.
+ * Copyright (c) Openizr. All Rights Reserved.
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  *
  */
 
-/* istanbul ignore file */
-
 import UIP from 'scripts/react/P';
+import UIIcon from 'scripts/react/Icon';
 import UILink from 'scripts/react/Link';
 import UITitle from 'scripts/react/Title';
 import UIImage from 'scripts/react/Image';
 import UIButton from 'scripts/react/Button';
+import UIOptions from 'scripts/react/Options';
 import markdown from 'scripts/helpers/markdown';
-import UIDropdown from 'scripts/react/Dropdown';
 import UITextarea from 'scripts/react/Textarea';
-import UICheckbox from 'scripts/react/Checkbox';
-import UIRadio from 'scripts/react/Radio';
 import UITextfield from 'scripts/react/Textfield';
+import UIFilePicker from 'scripts/react/FilePicker';
 import buildClass from 'scripts/helpers/buildClass';
-import UIFileUploader from 'scripts/react/FileUploader';
 import generateRandomId from 'scripts/helpers/generateRandomId';
 
 export {
   UIP,
+  UIIcon,
   UILink,
-  UIButton,
   UIImage,
-  UIDropdown,
   UITitle,
-  UITextfield,
+  UIButton,
+  UIOptions,
   UITextarea,
-  UICheckbox,
-  UIRadio,
-  UIFileUploader,
+  UITextfield,
+  UIFilePicker,
   markdown,
   buildClass,
   generateRandomId,
