@@ -1,5 +1,5 @@
 /**
- * Copyright (c) Matthieu Jabbour. All Rights Reserved.
+ * Copyright (c) Openizr. All Rights Reserved.
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
@@ -37,7 +37,7 @@ const getDimensions = (ratio: string): { width: number; height: number; } => {
     case 'panoramic':
       return { width: 16, height: 9 };
     default:
-      dimensions = ratio.split(':').map((value) => parseInt(value, 10));
+      dimensions = ratio.split('x').map((value) => parseInt(value, 10));
       return { width: dimensions[0], height: dimensions[1] };
   }
 };
@@ -45,14 +45,14 @@ const getDimensions = (ratio: string): { width: number; height: number; } => {
 /**
  * Image.
  */
-export default function UIImage(props: InferProps<typeof propTypes>): JSX.Element {
-  // eslint-disable-next-line object-curly-newline
-  const { id, src, alt, ratio, modifiers, itemProp } = props;
+function UIImage(props: InferProps<typeof propTypes>): JSX.Element {
+  const { id, src, alt } = props;
+  const { ratio, modifiers, itemProp } = props;
   const dimensions = getDimensions(ratio);
-  const className = buildClass('ui-image', `${ratio.replace(':', 'x')} ${modifiers}`.split(' '));
+  const className = buildClass('ui-image', `${ratio} ${modifiers}`);
 
   // Custom aspect ratio...
-  if (/^([0-9]+):([0-9]+)$/i.test(ratio) === true) {
+  if (/^([0-9]+)x([0-9]+)$/i.test(ratio)) {
     return (
       <img
         id={id as string}
@@ -62,21 +62,21 @@ export default function UIImage(props: InferProps<typeof propTypes>): JSX.Elemen
         className={className}
         width={dimensions.width}
         height={dimensions.height}
-        itemProp={(itemProp !== null) ? itemProp : undefined}
+        itemProp={itemProp as string}
       />
     );
   }
 
   // Standard aspect ratio...
   return (
-    <div className={className}>
+    <div id={id as string} className={className}>
       <img
         src={src}
         alt={alt}
         loading="lazy"
         width={dimensions.width}
         height={dimensions.height}
-        itemProp={(itemProp !== null) ? itemProp : undefined}
+        itemProp={itemProp as string}
       />
     </div>
   );
@@ -85,3 +85,5 @@ export default function UIImage(props: InferProps<typeof propTypes>): JSX.Elemen
 UIImage.displayName = 'UIImage';
 UIImage.propTypes = propTypes;
 UIImage.defaultProps = defaultProps;
+
+export default React.memo(UIImage);
