@@ -1,16 +1,24 @@
 /**
- * Copyright (c) Matthieu Jabbour. All Rights Reserved.
+ * Copyright (c) Openizr. All Rights Reserved.
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  *
  */
 
+/* eslint-disable react/no-danger, jsx-a11y/label-has-associated-control */
+
 import * as React from 'react';
+import UIIcon from 'scripts/react/Icon';
 import markdown from 'scripts/helpers/markdown';
 import PropTypes, { InferProps } from 'prop-types';
 import buildClass from 'scripts/helpers/buildClass';
 import generateRandomId from 'scripts/helpers/generateRandomId';
+
+type KeyType = 'default' | 'ctrlKey' | 'altKey' | 'shiftKey' | 'metaKey';
+type AllowedKeys = Record<KeyType, RegExp | null>;
+
+const JSXUIIcon = UIIcon as JSXElement;
 
 const propTypes = {
   id: PropTypes.string,
@@ -24,17 +32,26 @@ const propTypes = {
   helper: PropTypes.string,
   onBlur: PropTypes.func,
   onFocus: PropTypes.func,
+  onPaste: PropTypes.func,
   onChange: PropTypes.func,
+  onKeyDown: PropTypes.func,
   readonly: PropTypes.bool,
   transform: PropTypes.func,
   maxlength: PropTypes.number,
   modifiers: PropTypes.string,
   onIconClick: PropTypes.func,
   placeholder: PropTypes.string,
+  onIconKeyDown: PropTypes.func,
   autocomplete: PropTypes.string,
   name: PropTypes.string.isRequired,
   debounceTimeout: PropTypes.number,
-  allowedPattern: PropTypes.instanceOf(RegExp),
+  allowedKeys: PropTypes.shape({
+    altKey: PropTypes.instanceOf(RegExp),
+    metaKey: PropTypes.instanceOf(RegExp),
+    ctrlKey: PropTypes.instanceOf(RegExp),
+    default: PropTypes.instanceOf(RegExp),
+    shiftKey: PropTypes.instanceOf(RegExp),
+  }),
   iconPosition: PropTypes.oneOf(['left', 'right']),
   type: PropTypes.oneOf(['text', 'email', 'number', 'password', 'search', 'tel', 'url']),
 };
@@ -53,101 +70,150 @@ const defaultProps = {
   modifiers: '',
   onBlur: null,
   onFocus: null,
+  onPaste: null,
   onChange: null,
+  onKeyDown: null,
+  allowedKeys: {},
+  transform: null,
   readonly: false,
   maxlength: null,
+  onIconClick: null,
   placeholder: null,
   autocomplete: 'on',
   iconPosition: 'left',
-  debounceTimeout: null,
-  allowedPattern: null,
-  onIconClick: undefined,
-  transform: (value: string): string => value,
+  onIconKeyDown: null,
+  debounceTimeout: 0,
 };
 
-const specialKeysRegexp = /(Tab|Backspace|Delete|Enter|ArrowRight|ArrowLeft|ArrowDown|ArrowUp)/;
+const keyTypes: KeyType[] = ['default', 'ctrlKey', 'altKey', 'shiftKey', 'metaKey'];
+const specialKeysRegexp = /Tab|Control|Shift|Meta|ContextMenu|Alt|Escape|Insert|Home|End|AltGraph|NumLock|Backspace|Delete|Enter|ArrowRight|ArrowLeft|ArrowDown|ArrowUp/;
+const defaultTransform = (value: string): string[] => [value];
 
 /**
- * Textfield.
+ * Text field.
  */
-export default function UITextfield(props: InferProps<typeof propTypes>): JSX.Element {
-  const {
-    id, modifiers, label, helper, onChange, value, name, readonly, step, onIconClick, autocomplete,
-    placeholder, iconPosition, icon, onBlur, type, size, max, min, maxlength, onFocus,
-    debounceTimeout, allowedPattern,
-  } = props;
+function UITextfield(props: InferProps<typeof propTypes>): JSX.Element {
+  const { type, size, max } = props;
+  const { name, readonly, step } = props;
+  const { id, modifiers, label } = props;
+  const { helper, onChange, value } = props;
+  const { min, maxlength, onFocus } = props;
+  const { iconPosition, icon, onBlur } = props;
+  const { onIconKeyDown, transform, onPaste } = props;
+  const { onIconClick, autocomplete, placeholder } = props;
+  const { debounceTimeout, allowedKeys, onKeyDown } = props;
   const [randomId] = React.useState(generateRandomId);
+  const actualTransform = transform || defaultTransform;
   const inputRef = React.useRef<HTMLInputElement>(null);
-  const [timeout, setTimeout] = React.useState<number | null>(null);
-  const { transform } = (props as { transform: (value?: string | null) => string });
-  const [currentValue, setCurrentValue] = React.useState(transform(value));
-  const [cursorPosition, setCursorPosition] = React.useState<number | null>(null);
+  const timeout = React.useRef<NodeJS.Timeout | null>(null);
   const isDisabled = (modifiers as string).includes('disabled');
-  const className = buildClass('ui-textfield', (modifiers as string).split(' '));
-  const globalAllowedPattern = (allowedPattern instanceof RegExp) ? new RegExp(allowedPattern.source, 'g') : null;
+  const className = buildClass('ui-textfield', modifiers as string);
+  const [cursorPosition, setCursorPosition] = React.useState<number | null>(null);
+  const [currentValue, setCurrentValue] = React.useState(() => actualTransform(value, 0)[0]);
 
-  const changeValue = (event: React.ChangeEvent<HTMLInputElement>): void => {
-    const filteredValue = (globalAllowedPattern !== null)
-      ? (event.target.value.match(globalAllowedPattern) || []).join('')
+  // Memoïzes global version of allowed keys RegExps (required for filtering out a whole text).
+  const globalAllowedKeys = React.useMemo(() => keyTypes.reduce((allAllowedKeys, keyType) => {
+    const allowedKeysForType = (allowedKeys as AllowedKeys)[keyType];
+    return {
+      ...allAllowedKeys,
+      [keyType]: (allowedKeysForType !== undefined && allowedKeysForType !== null)
+        ? new RegExp(allowedKeysForType.source, `${allowedKeysForType.flags}g`)
+        : null,
+    };
+  }, {} as AllowedKeys), [allowedKeys]);
+
+  // -----------------------------------------------------------------------------------------------
+  // CALLBACKS DECLARATION.
+  // -----------------------------------------------------------------------------------------------
+
+  const handleChange = (event: React.ChangeEvent<HTMLInputElement>, filter = true): void => {
+    const filteredValue = (filter && globalAllowedKeys.default !== null)
+      ? (event.target.value.match(globalAllowedKeys.default) || []).join('')
       : event.target.value;
-    const newValue = transform(filteredValue);
+    const { selectionStart } = event.target;
+    const [newValue, newCursorPosition] = actualTransform(filteredValue, selectionStart);
     setCurrentValue(newValue);
-    // At this point, the input's value has already changed, which means the cursor's position is
-    // at n + 1, which is why we substract 1 when checking last position.
-    const currentCursorPosition = event.target.selectionStart as number;
-    const isAtTheEnd = currentCursorPosition - 1 >= currentValue.length;
-    setCursorPosition(isAtTheEnd ? newValue.length : currentCursorPosition);
+    if (newCursorPosition !== undefined) {
+      setCursorPosition(newCursorPosition);
+    } else {
+      // At this point, the input's value has already changed, which means the cursor's position is
+      // at n + 1, which is why we substract 1 when checking last position.
+      const currentCursorPosition = event.target.selectionStart as number;
+      const isAtTheEnd = currentCursorPosition - 1 >= currentValue.length;
+      setCursorPosition(isAtTheEnd ? newValue.length : currentCursorPosition);
+    }
     if (onChange !== undefined && onChange !== null) {
-      window.clearTimeout(timeout as number);
-      // This debounce system prevents triggering `onChange` hooks too many times when user is
+      window.clearTimeout(timeout.current as NodeJS.Timeout);
+      // This debounce system prevents triggering `onChange` callback too many times when user is
       // still typing to improve performance and make UI more reactive on low-perfomance devices.
-      setTimeout(window.setTimeout(() => {
-        onChange(newValue);
-      }, debounceTimeout || 0));
+      timeout.current = setTimeout(() => {
+        onChange(newValue, event);
+      }, debounceTimeout as number);
     }
   };
 
-  const checkPattern = (event: React.KeyboardEvent<HTMLInputElement>): void => {
+  const handleKeyDown = (event: React.KeyboardEvent<HTMLInputElement>): void => {
+    let allowedKeysForEvent = (allowedKeys as AllowedKeys).default || null;
+    if (event.ctrlKey === true) {
+      allowedKeysForEvent = (allowedKeys as AllowedKeys).ctrlKey || null;
+    } else if (event.shiftKey === true) {
+      allowedKeysForEvent = (allowedKeys as AllowedKeys).shiftKey || null;
+    } else if (event.altKey === true) {
+      allowedKeysForEvent = (allowedKeys as AllowedKeys).altKey || null;
+    } else if (event.metaKey === true) {
+      allowedKeysForEvent = (allowedKeys as AllowedKeys).metaKey || null;
+    }
     if (
-      allowedPattern instanceof RegExp
-      && !allowedPattern.test(event.key)
+      allowedKeysForEvent !== null
+      && !allowedKeysForEvent.test(event.key)
       && !specialKeysRegexp.test(event.key)
-      && !event.ctrlKey
     ) {
       event.preventDefault();
+    } else if (onKeyDown) {
+      onKeyDown(event);
     }
   };
 
-  const paste = (event: React.ClipboardEvent<HTMLInputElement>): void => {
-    const cursor = (event.target as HTMLInputElement).selectionStart as number;
-    const filteredValue = (globalAllowedPattern !== null)
-      ? (event.clipboardData.getData('text').match(globalAllowedPattern) || []).join('')
+  const handlePaste = (event: React.ClipboardEvent<HTMLInputElement>): void => {
+    const selectionEnd = (event.target as HTMLInputElement).selectionEnd as number;
+    const selectionStart = (event.target as HTMLInputElement).selectionStart as number;
+    const filteredValue = (globalAllowedKeys.default !== null)
+      ? (event.clipboardData.getData('text').match(globalAllowedKeys.default) || []).join('')
       : event.clipboardData.getData('text');
-    changeValue({
+    handleChange({
       target: {
-        value: `${currentValue.slice(0, cursor)}${filteredValue}${currentValue.slice(cursor)}`,
-        selectionStart: cursor + filteredValue.length,
+        value: `${currentValue.slice(0, selectionStart)}${filteredValue}${currentValue.slice(selectionEnd)}`,
+        selectionStart: selectionStart + filteredValue.length,
       },
-    } as unknown as React.ChangeEvent<HTMLInputElement>);
+    } as unknown as React.ChangeEvent<HTMLInputElement>, false);
     event.preventDefault();
+    if (onPaste !== undefined && onPaste !== null) {
+      onPaste(event);
+    }
   };
 
-  const blurField = (): void => {
+  const handleBlur = (event: React.FocusEvent<HTMLInputElement>): void => {
     if (onBlur !== undefined && onBlur !== null) {
-      onBlur(currentValue);
+      onBlur(currentValue, event);
     }
   };
 
-  const focusField = (): void => {
+  const handleFocus = (event: React.FocusEvent<HTMLInputElement>): void => {
     if (onFocus !== undefined && onFocus !== null) {
-      onFocus(currentValue);
+      onFocus(currentValue, event);
     }
   };
 
-  // Updates current value each time the `value` property is changed.
+  // -----------------------------------------------------------------------------------------------
+  // PROPS REACTIVITY MANAGEMENT.
+  // -----------------------------------------------------------------------------------------------
+
+  // Updates current value whenever `value` and `transform` props change.
   React.useEffect(() => {
-    setCurrentValue(transform(value));
-  }, [value]);
+    const [newValue, newCursorPosition] = actualTransform(value, 0);
+    setCurrentValue(newValue);
+    setCursorPosition(newCursorPosition);
+  }, [value, actualTransform]);
 
   // Re-positions cursor at the right place when using transform function.
   React.useEffect(() => {
@@ -155,35 +221,49 @@ export default function UITextfield(props: InferProps<typeof propTypes>): JSX.El
       (inputRef.current as HTMLInputElement).selectionStart = cursorPosition;
       (inputRef.current as HTMLInputElement).selectionEnd = cursorPosition;
     }
-  }, [currentValue, cursorPosition]);
+  }, [currentValue, cursorPosition, type]);
+
+  // -----------------------------------------------------------------------------------------------
+  // COMPONENT RENDERING.
+  // -----------------------------------------------------------------------------------------------
 
   const children = [
-    (icon !== null)
-      // eslint-disable-next-line jsx-a11y/click-events-have-key-events
-      ? <i key="icon" role="button" tabIndex={0} className="ui-textfield__wrapper__icon" onClick={onIconClick || undefined}>{icon}</i>
+    (icon !== null && icon !== undefined)
+      ? (
+        <span
+          key="icon"
+          tabIndex={0}
+          role="button"
+          onClick={onIconClick as undefined}
+          onKeyDown={onIconKeyDown as undefined}
+          className="ui-textfield__wrapper__icon"
+        >
+          <JSXUIIcon name={icon} />
+        </span>
+      )
       : null,
     <input
       key="input"
       name={name}
       id={randomId}
       ref={inputRef}
-      onPaste={paste}
-      onBlur={blurField}
-      onFocus={focusField}
-      onKeyDown={checkPattern}
       max={max as number}
       min={min as number}
       step={step as number}
       type={type as string}
       size={size as number}
       disabled={isDisabled}
+      onBlur={handleBlur}
+      onFocus={handleFocus}
       value={currentValue as string}
       readOnly={readonly as boolean}
       maxLength={maxlength as number}
       autoComplete={autocomplete as string}
       placeholder={placeholder as string}
       className="ui-textfield__wrapper__field"
-      onChange={(readonly === false && !isDisabled) ? changeValue : undefined}
+      onPaste={(readonly === false && !isDisabled) ? handlePaste : undefined}
+      onChange={(readonly === false && !isDisabled) ? handleChange : undefined}
+      onKeyDown={(readonly === false && !isDisabled) ? handleKeyDown : undefined}
     />,
   ];
 
@@ -193,13 +273,14 @@ export default function UITextfield(props: InferProps<typeof propTypes>): JSX.El
       className={className}
     >
       {(label !== null && label !== undefined)
-        // eslint-disable-next-line react/no-danger, jsx-a11y/label-has-associated-control
         ? <label className="ui-textfield__label" htmlFor={randomId} dangerouslySetInnerHTML={{ __html: markdown(label) }} />
         : null}
       <div className="ui-textfield__wrapper">
         {(iconPosition === 'left') ? children : children.reverse()}
       </div>
-      {(helper !== null) ? <span className="ui-textfield__helper">{helper}</span> : null}
+      {(helper !== null && helper !== undefined)
+        ? <span className="ui-textfield__helper" dangerouslySetInnerHTML={{ __html: markdown(helper) }} />
+        : null}
     </div>
   );
 }
@@ -207,3 +288,5 @@ export default function UITextfield(props: InferProps<typeof propTypes>): JSX.El
 UITextfield.propTypes = propTypes;
 UITextfield.defaultProps = defaultProps;
 UITextfield.displayName = 'UITextfield';
+
+export default React.memo(UITextfield);
