@@ -1,5 +1,5 @@
 /**
- * Copyright (c) Matthieu Jabbour. All Rights Reserved.
+ * Copyright (c) Openizr. All Rights Reserved.
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
@@ -25,31 +25,32 @@ const defaultProps = {
   id: null,
   level: '1',
   modifiers: '',
-  itemProp: undefined,
+  itemProp: null,
 };
 
 /**
  * Title.
  */
-export default function UITitle(props: InferProps<typeof propTypes>): JSX.Element {
-  // eslint-disable-next-line object-curly-newline
-  const { id, level, label, modifiers, itemProp } = props;
-  const modifiersList = (modifiers as string).split(' ');
+function UITitle(props: InferProps<typeof propTypes>): JSX.Element {
+  const { level, modifiers, itemProp } = props;
+  let modifiersList = props.modifiers as string;
 
   // Checks if any of the given modifiers corresponds to a valid level (1, 2, ...).
   // By default, if no level is specified in modifiers, we set it to the `level` prop.
   if (/(^|\s)([1-6])($|\s)/i.test(modifiers as string) === false) {
-    modifiersList.push(level as string);
+    modifiersList += ` ${level as string}`;
   }
 
   return React.createElement(`h${level}`, {
-    id,
+    id: props.id,
     itemProp,
     className: buildClass('ui-title', modifiersList),
-    dangerouslySetInnerHTML: { __html: markdown(label) },
-  });
+    dangerouslySetInnerHTML: { __html: markdown(props.label) },
+  }) as JSXElement;
 }
 
 UITitle.propTypes = propTypes;
 UITitle.defaultProps = defaultProps;
 UITitle.displayName = 'UITitle';
+
+export default React.memo(UITitle);
