@@ -1,5 +1,9 @@
 /**
- * Copyright (c) Matthieu Jabbour. All Rights Reserved.
+ * @jest-environment jsdom
+ */
+
+/**
+ * Copyright (c) Openizr. All Rights Reserved.
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
@@ -8,63 +12,42 @@
 
 import React from 'react';
 import UIImage from 'scripts/react/Image';
-import { act } from 'react-dom/test-utils';
-import { render, unmountComponentAtNode } from 'react-dom';
+import { render } from '@testing-library/react';
 
-let container = document.createElement('div');
+const JSXUIImage = UIImage as JSXElement;
 
 describe('react/UIImage', () => {
   beforeEach(() => {
-    container = document.createElement('div');
-    document.body.appendChild(container);
     jest.clearAllMocks();
   });
 
-  afterEach(() => {
-    unmountComponentAtNode(container);
-    container.remove();
-    ((container as unknown) as null) = null;
-  });
-
   test('renders correctly - basic', () => {
-    act(() => {
-      render(<UIImage alt="test" src="https://test.com/a.jpg" ratio="square" modifiers="large" />, container);
-    });
-    expect(container).toMatchSnapshot();
+    const { container } = render(<JSXUIImage alt="test" src="https://test.com/a.jpg" ratio="square" modifiers="large" />);
+    expect(container.firstChild).toMatchSnapshot();
   });
 
   test('renders correctly - custom ratio', () => {
-    act(() => {
-      render(<UIImage alt="test" src="https://test.com/a.jpg" ratio="25:32" />, container);
-    });
-    expect(container).toMatchSnapshot();
+    const { container } = render(<JSXUIImage alt="test" src="https://test.com/a.jpg" ratio="25x32" />);
+    expect(container.firstChild).toMatchSnapshot();
   });
 
   test('renders correctly - with id', () => {
-    act(() => {
-      render(<UIImage id="test" alt="test" src="https://test.com/a.jpg" ratio="portrait" />, container);
-    });
-    expect(container).toMatchSnapshot();
+    const { container } = render(<JSXUIImage id="test" alt="test" src="https://test.com/a.jpg" ratio="portrait" />);
+    expect(container.firstChild).toMatchSnapshot();
   });
 
   test('renders correctly - standard ratio with itemprop', () => {
-    act(() => {
-      render(<UIImage itemProp="image" alt="test" src="https://test.com/a.jpg" ratio="landscape" />, container);
-    });
-    expect(container).toMatchSnapshot();
+    const { container } = render(<JSXUIImage itemProp="image" alt="test" src="https://test.com/a.jpg" ratio="landscape" />);
+    expect(container.firstChild).toMatchSnapshot();
   });
 
   test('renders correctly - custom ratio with itemprop', () => {
-    act(() => {
-      render(<UIImage itemProp="image" alt="test" src="https://test.com/a.jpg" ratio="1:5" />, container);
-    });
-    expect(container).toMatchSnapshot();
+    const { container } = render(<JSXUIImage itemProp="image" alt="test" src="https://test.com/a.jpg" ratio="1x5" />);
+    expect(container.firstChild).toMatchSnapshot();
   });
 
   test('renders correctly - with panoramic ratio', () => {
-    act(() => {
-      render(<UIImage alt="test" src="https://test.com/a.jpg" ratio="panoramic" />, container);
-    });
-    expect(container).toMatchSnapshot();
+    const { container } = render(<JSXUIImage alt="test" src="https://test.com/a.jpg" ratio="panoramic" />);
+    expect(container.firstChild).toMatchSnapshot();
   });
 });
