@@ -1,40 +1,36 @@
 /**
- * Copyright (c) Matthieu Jabbour. All Rights Reserved.
+ * Copyright (c) Openizr. All Rights Reserved.
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  *
  */
 
-/* istanbul ignore file */
-
 import UIP from 'scripts/vue/P.vue';
+import UIIcon from 'scripts/vue/Icon.vue';
 import UILink from 'scripts/vue/Link.vue';
 import UITitle from 'scripts/vue/Title.vue';
 import UIImage from 'scripts/vue/Image.vue';
-import UIRadio from 'scripts/vue/Radio.vue';
 import UIButton from 'scripts/vue/Button.vue';
-import UIDropdown from 'scripts/vue/Dropdown.vue';
-import UITextarea from 'scripts/vue/Textarea.vue';
-import UICheckbox from 'scripts/vue/Checkbox.vue';
+import UIOptions from 'scripts/vue/Options.vue';
 import markdown from 'scripts/helpers/markdown';
+import UITextarea from 'scripts/vue/Textarea.vue';
 import UITextfield from 'scripts/vue/Textfield.vue';
 import buildClass from 'scripts/helpers/buildClass';
-import UIFileUploader from 'scripts/vue/FileUploader.vue';
+import UIFilePicker from 'scripts/vue/FilePicker.vue';
 import generateRandomId from 'scripts/helpers/generateRandomId';
 
 export {
   UIP,
   UILink,
-  UIButton,
-  UIImage,
-  UIDropdown,
+  UIIcon,
   UITitle,
+  UIImage,
+  UIButton,
+  UIOptions,
   UITextfield,
   UITextarea,
-  UICheckbox,
-  UIRadio,
-  UIFileUploader,
+  UIFilePicker,
   markdown,
   buildClass,
   generateRandomId,
