@@ -1,13 +1,17 @@
 /**
- * Copyright (c) Matthieu Jabbour. All Rights Reserved.
+ * @jest-environment jsdom
+ */
+
+/**
+ * Copyright (c) Openizr. All Rights Reserved.
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  *
  */
 
-import { mount } from '@vue/test-utils';
 import UIImage from 'scripts/vue/Image.vue';
+import { render } from '@testing-library/vue';
 
 describe('vue/UIImage', () => {
   beforeEach(() => {
@@ -15,52 +19,48 @@ describe('vue/UIImage', () => {
   });
 
   test('renders correctly - basic', () => {
-    const wrapper = mount(UIImage, {
-      propsData: {
+    const { container } = render(UIImage, {
+      props: {
         alt: 'test', ratio: 'square', src: 'https://test.com/a.jpg', modifiers: 'large',
       },
     });
-    expect(wrapper.html()).toMatchSnapshot();
+    expect(container.firstChild).toMatchSnapshot();
   });
 
   test('renders correctly - custom ratio', () => {
-    const wrapper = mount(UIImage, {
-      propsData: { alt: 'test', ratio: '25:32', src: 'https://test.com/a.jpg' },
-    });
-    expect(wrapper.html()).toMatchSnapshot();
+    const { container } = render(UIImage, { props: { alt: 'test', ratio: '25x32', src: 'https://test.com/a.jpg' } });
+    expect(container.firstChild).toMatchSnapshot();
   });
 
   test('renders correctly - with id', () => {
-    const wrapper = mount(UIImage, {
-      propsData: {
+    const { container } = render(UIImage, {
+      props: {
         id: 'test', alt: 'test', ratio: 'portrait', src: 'https://test.com/a.jpg',
       },
     });
-    expect(wrapper.html()).toMatchSnapshot();
+    expect(container.firstChild).toMatchSnapshot();
   });
 
   test('renders correctly - standard ratio with itemprop', () => {
-    const wrapper = mount(UIImage, {
-      propsData: {
+    const { container } = render(UIImage, {
+      props: {
         itemProp: 'image', alt: 'test', ratio: 'landscape', src: 'https://test.com/a.jpg',
       },
     });
-    expect(wrapper.html()).toMatchSnapshot();
+    expect(container.firstChild).toMatchSnapshot();
   });
 
   test('renders correctly - custom ratio with itemprop', () => {
-    const wrapper = mount(UIImage, {
-      propsData: {
-        itemProp: 'image', alt: 'test', ratio: '1:5', src: 'https://test.com/a.jpg',
+    const { container } = render(UIImage, {
+      props: {
+        itemProp: 'image', alt: 'test', ratio: '1x5', src: 'https://test.com/a.jpg',
       },
     });
-    expect(wrapper.html()).toMatchSnapshot();
+    expect(container.firstChild).toMatchSnapshot();
   });
 
   test('renders correctly - with panoramic ratio', () => {
-    const wrapper = mount(UIImage, {
-      propsData: { alt: 'test', ratio: 'panoramic', src: 'https://test.com/a.jpg' },
-    });
-    expect(wrapper.html()).toMatchSnapshot();
+    const { container } = render(UIImage, { props: { alt: 'test', ratio: 'panoramic', src: 'https://test.com/a.jpg' } });
+    expect(container.firstChild).toMatchSnapshot();
   });
 });
