@@ -1,13 +1,17 @@
 /**
- * Copyright (c) Matthieu Jabbour. All Rights Reserved.
+ * @jest-environment jsdom
+ */
+
+/**
+ * Copyright (c) Openizr. All Rights Reserved.
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  *
  */
 
-import { mount } from '@vue/test-utils';
 import UITitle from 'scripts/vue/Title.vue';
+import { render } from '@testing-library/vue';
 
 describe('vue/UITitle', () => {
   beforeEach(() => {
@@ -15,30 +19,22 @@ describe('vue/UITitle', () => {
   });
 
   test('renders correctly - basic', () => {
-    const wrapper = mount(UITitle, {
-      propsData: { level: '1', modifiers: 'large' },
-    });
-    expect(wrapper.html()).toMatchSnapshot();
+    const { container } = render(UITitle, { props: { label: 'Test', modifiers: 'large' } });
+    expect(container.firstChild).toMatchSnapshot();
   });
 
   test('renders correctly - modifier different than level', () => {
-    const wrapper = mount(UITitle, {
-      propsData: { label: 'Test', level: '1', modifiers: '5 large' },
-    });
-    expect(wrapper.html()).toMatchSnapshot();
+    const { container } = render(UITitle, { props: { label: 'Test', level: '1', modifiers: '5 large' } });
+    expect(container.firstChild).toMatchSnapshot();
   });
 
   test('renders correctly - with id', () => {
-    const wrapper = mount(UITitle, {
-      propsData: { label: 'Test', id: 'test' },
-    });
-    expect(wrapper.html()).toMatchSnapshot();
+    const { container } = render(UITitle, { props: { label: 'Test', id: 'test' } });
+    expect(container.firstChild).toMatchSnapshot();
   });
 
   test('renders correctly - with itemprop', () => {
-    const wrapper = mount(UITitle, {
-      propsData: { label: '*Test*', itemProp: 'name' },
-    });
-    expect(wrapper.html()).toMatchSnapshot();
+    const { container } = render(UITitle, { props: { label: '*Test*', itemProp: 'name' } });
+    expect(container.firstChild).toMatchSnapshot();
   });
 });
