@@ -1,5 +1,5 @@
 /**
- * Copyright (c) Matthieu Jabbour. All Rights Reserved.
+ * Copyright (c) Openizr. All Rights Reserved.
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
@@ -19,23 +19,17 @@ export default function Home(): JSX.Element {
         <br />
         <a href="/images" className="cols-3">IMAGES</a>
         <br />
-        <a href="/radios" className="cols-3">RADIOS</a>
-        <br />
         <a href="/buttons" className="cols-3">BUTTONS</a>
         <br />
         <a href="/markdown" className="cols-3">MARKDOWN</a>
         <br />
-        <a href="/dropdowns" className="cols-3">DROPDOWNS</a>
+        <a href="/options" className="cols-3">OPTIONS</a>
         <br />
         <a href="/textfields" className="cols-3">TEXTFIELDS</a>
         <br />
         <a href="/textareas" className="cols-3">TEXTAREAS</a>
         <br />
-        <a href="/file-uploaders" className="cols-3">FILE UPLOADERS</a>
-        <br />
-        <a href="/radios" className="cols-3">RADIOS</a>
-        <br />
-        <a href="/checkboxes" className="cols-3">CHECKBOXES</a>
+        <a href="/file-pickers" className="cols-3">FILE PICKERS</a>
         <br />
         <a href="/icons" className="cols-3">ICONS</a>
       </main>
