@@ -1,5 +1,5 @@
 /**
- * Copyright (c) Matthieu Jabbour. All Rights Reserved.
+ * Copyright (c) Openizr. All Rights Reserved.
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
@@ -7,25 +7,25 @@
  */
 
 import * as React from 'react';
-import { UIImage } from 'sonar-ui/react';
+import { UIImage } from 'biuty/react';
 
 /**
  * Images page.
  */
 export default function Images(): JSX.Element {
   return (
-    <div className="vgap-5">
-      <main className="ui-page ui-block cols-3 hgap-3 vgap-5">
+    <div>
+      <main className="grid cols-3 hgap-3 vgap-5 items-center">
         <a href="/" className="cols-3">GO BACK</a>
         <UIImage ratio="square" src="https://picsum.photos/200/600" alt="img" />
         <UIImage ratio="square" src="https://picsum.photos/300/300" alt="img" />
         <UIImage ratio="square" src="https://picsum.photos/600/200" alt="img" />
-        <UIImage ratio="square" modifiers="large" src="https://picsum.photos/200/600" alt="img" />
-        <UIImage ratio="square" modifiers="large" src="https://picsum.photos/300/300" alt="img" />
-        <UIImage ratio="square" modifiers="large" src="https://picsum.photos/600/200" alt="img" />
-        <UIImage ratio="square" modifiers="small" src="https://picsum.photos/200/600" alt="img" />
-        <UIImage ratio="square" modifiers="small" src="https://picsum.photos/300/300" alt="img" />
-        <UIImage ratio="square" modifiers="small" src="https://picsum.photos/600/200" alt="img" />
+        <UIImage ratio="square" src="https://picsum.photos/200/600" alt="img" />
+        <UIImage ratio="150x250" src="https://picsum.photos/300/300" alt="img" />
+        <UIImage ratio="panoramic" src="https://picsum.photos/600/200" alt="img" />
+        <UIImage ratio="landscape" src="https://picsum.photos/200/600" alt="img" />
+        <UIImage ratio="portrait" src="https://picsum.photos/300/300" alt="img" />
+        <UIImage ratio="portrait" src="https://picsum.photos/600/200" alt="img" />
       </main>
     </div>
   );
