@@ -1,13 +1,20 @@
 /**
- * Copyright (c) Matthieu Jabbour. All Rights Reserved.
+ * Copyright (c) Openizr. All Rights Reserved.
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  *
  */
 
-import * as React from 'react';
-import { UITitle, UIP, UILink } from 'sonar-ui/react';
+import 'biuty';
+import React from 'react';
+import { UIP, UITitle, UILink } from 'biuty/react';
+
+const { log } = console;
+const onClick = (event: React.MouseEvent<HTMLAnchorElement>): void => {
+  log('Clicked!', event);
+  event.preventDefault();
+};
 
 /**
  * Typography page.
@@ -24,11 +31,14 @@ export default function Typography(): JSX.Element {
         <UITitle level="5" label={'ui-title ui-title--5\nui-title ui-title--5'} />
         <UITitle level="6" label={'ui-title ui-title--6\nui-title ui-title--6'} />
         <UIP label="ui-p ui-p ui-p ui-p ~ui-p italic~ _ui-p underline_ **ui-p strong** *ui-p emphasis* ui-p ui-p ui-p ui-p ui-p ui-p ui-p ui-p ui-p ui-p ui-p ui-p ui-p ui-p ui-p ui-p ui-p ui-p ui-p ui-p ui-p ui-p ui-p ui-p ui-p ui-p ui-p " />
-        <div>
-          <div className="ui-button-group">
-            <UILink href="/" target="_blank" rel="nofollow noopener noreferer" title="link title" label="ui-link" />
-          </div>
-        </div>
+        <UILink
+          href="/"
+          target="_blank"
+          rel="nofollow noopener noreferer"
+          title="link title"
+          label="ui-link"
+          onClick={onClick}
+        />
       </main>
     </div>
   );
