@@ -1,5 +1,5 @@
 /**
- * Copyright (c) Matthieu Jabbour. All Rights Reserved.
+ * Copyright (c) Openizr. All Rights Reserved.
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
@@ -8,13 +8,11 @@
 
 export default {
   '/': () => import('scripts/pages/Home.vue'),
-  '/buttons': () => import('scripts/pages/Buttons.vue'),
+  '/buttons': () => import('scripts/pages/ButtonsPage.vue'),
   '/images': () => import('scripts/pages/Images.vue'),
-  '/dropdowns': () => import('scripts/pages/Dropdowns.vue'),
+  '/options': () => import('scripts/pages/OptionsPage.vue'),
   '/textfields': () => import('scripts/pages/Textfields.vue'),
   '/file-uploaders': () => import('scripts/pages/FileUploaders.vue'),
   '/textareas': () => import('scripts/pages/Textareas.vue'),
-  '/radios': () => import('scripts/pages/Radios.vue'),
-  '/checkboxes': () => import('scripts/pages/Checkboxes.vue'),
   '/typography': () => import('scripts/pages/Typography.vue'),
 } as Record<string, unknown>;
